requirements.txt
Django==1.11
pkg-resources==0.0.0
python-dateutil==2.7.3
pytz==2018.4
six==1.11.0
